% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_age.R
\name{sample_age}
\alias{sample_age}
\title{Create fictional ages based on birth date and survey date}
\usage{
sample_age(start_date, end_date = Sys.Date(), error = 0)
}
\arguments{
\item{start_date}{date or vector of dates; birth date or other origin date}

\item{end_date}{dateor vector of dates; survey date or other completion date}

\item{error}{numeric; amount of error to include, from 0 (none) to 100
(full dataset)}
}
\description{
Create fictional ages based on birth date and survey date
}
\examples{
sample_age(start_date = "1900-01-01", end_date = "1960-07-31", error = 100)

}
