% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_ip.R
\name{sample_ip}
\alias{sample_ip}
\title{Create fictional IP addresses}
\usage{
sample_ip(count)
}
\arguments{
\item{count}{integer; number of responses}
}
\description{
Create fictional IP addresses
}
\examples{
sample_ip(count = 5)

}
