% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_many.R
\name{sample_many}
\alias{sample_many}
\title{Calculate multiple booleans}
\usage{
sample_many(var, d, count = 100)
}
\arguments{
\item{var}{string; variable name}

\item{d}{dataset; contains labels and probabilities}

\item{count}{integer; number of responses (rows)}
}
\description{
This function provides up to the maximum allowed number of responses. Some
responses will have less and if your "miss_pct" is greater than zero, some
responses will have no selections.
}
\examples{
d <- data.frame(variable = "my_letters", type = "select-many", miss_pct = 10,
                options = letters[1:5], labels = letters[1:5], max_opts = 3,
                probability_1 = 9:5, probability_2 = 1:5)

sample_many(var = "my_letters", d = d, count = 10)

}
