% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_one.R
\name{sample_one}
\alias{sample_one}
\title{Create a random vector of responses}
\usage{
sample_one(options, probs, count = 100, miss_pct = 10)
}
\arguments{
\item{options}{vector of strings; value options}

\item{probs}{vector of numbers; probability options}

\item{count}{integer; number of responses (rows)

Create a vector of values based on pre-defined choices and probabilities.
You can also program in a certain percent of missing values.}

\item{miss_pct}{number between 0 and 100; percent of responses expected
to be missing}
}
\description{
Create a random vector of responses
}
\examples{
sample_one(options = letters[1:3], probs = 1:3, count = 10, miss_pct = 20)

}
