% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_error.R
\name{sampling_error}
\alias{sampling_error}
\title{Add sampling data errors}
\usage{
sampling_error(var, type = "numeric", error = 5)
}
\arguments{
\item{var}{vector of values to be modified}

\item{type}{type of data in the vector: accepts "character" or "numeric"}

\item{error}{numeric; amount of error to include, from 0 (none) to 100
(full dataset)}
}
\description{
Add sampling data errors
}
\examples{
d <- data.frame(x = signif(rnorm(10, mean = 0, sd = 100), digits = 4))
d$y <- sampling_error(d$x, error = 50)

}
