% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_categorical_set.R
\name{select_categorical_set}
\alias{select_categorical_set}
\title{Create a set of single-selection categorical variables}
\usage{
select_categorical_set(d, count)
}
\arguments{
\item{d}{dataset containing variable details}

\item{count}{integer; number of responses (rows)}
}
\description{
Create a set of single-selection categorical variables
}
\examples{
d <- data.frame(
  type = "select-one",
  variable = "my_letters",
  options = letters[1:5],
  probability_1 = 1:5,
  miss_pct = 10
)

select_categorical_set(d, count = 10)

}
