% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_continuous_set.R
\name{select_continuous_set}
\alias{select_continuous_set}
\title{Create a set of single-selection continuous variables}
\usage{
select_continuous_set(d, count)
}
\arguments{
\item{d}{dataset containing variable details}

\item{count}{integer; number of responses (rows)}
}
\description{
Create a set of single-selection continuous variables
}
\examples{
d <- data.frame(type = "normal", variable = "my_numbers", miss_pct = 10,
                min_val = 0,	max_val = 100, mean_val = 75,	sd_val = 10)

select_continuous_set(d, count = 10)

}
