% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_dates_set.R
\name{select_dates_set}
\alias{select_dates_set}
\title{Create a set of single-selection date variables}
\usage{
select_dates_set(d, count)
}
\arguments{
\item{d}{dataset containing variable details}

\item{count}{integer; number of responses (rows)}
}
\description{
Create a set of single-selection date variables
}
\examples{
d <- data.frame(variable = "my_dates", type = "date", miss_pct = 10,
                min_val = "6/1/2025",	max_val = "8/31/2025")

select_dates_set(d, count = 10)

}
