% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_many_set.R
\name{select_many_set}
\alias{select_many_set}
\title{Create a set of multiple-selection variables}
\usage{
select_many_set(d, count = 100)
}
\arguments{
\item{d}{dataset containing variable details}

\item{count}{integer; number of responses (rows)}
}
\description{
Create a set of multiple-selection variables
}
\examples{
d <- data.frame(variable = "my_letters", type = "select-many", miss_pct = 10,
                options = letters[1:5], labels = letters[1:5], max_opts = 3,
                probability_1 = 9:5, probability_2 = 1:5)

select_many_set(d, count = 10)

}
