% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toysurveydata-package.R
\docType{package}
\name{toysurveydata-package}
\alias{toysurveydata}
\alias{toysurveydata-package}
\title{Toy Survey Data}
\description{
This package contains funtions to create fake survey data. It contains the following functions,
which were coded to simulate Community Health Assessment datasets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ajstamm/toy_survey_data}
  \item \url{https://ajstamm.github.io/toysurveydata/}
  \item Report bugs at \url{https://github.com/ajstamm/toy_survey_data/issues}
}

}
\author{
\strong{Maintainer}: Abigail Stamm \email{abby.stamm@state.mn.us} (author of R package)

Authors:
\itemize{
  \item Abigail Stamm \email{abby.stamm@state.mn.us} (author of R package)
}

}
\keyword{internal}
