% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_error.R
\name{sampling_error}
\alias{sampling_error}
\title{Add sampling data errors}
\usage{
sampling_error(var, type = "numeric")
}
\arguments{
\item{var}{Vector of values to be modified}

\item{type}{Type of data in the vector: accepts "character" or "numeric"}
}
\description{
Add sampling data errors
}
\examples{
d <- data.frame(x = signif(rnorm(10000, mean = 0, sd = 100), digits = 4))
d$y <- sampling_error(d$x)

}
